260105 @ kim b j : 수온정보 그래프 지난 작업분 반영

+++ app/src/main/java/com/bj/fishingdiary/domain/entity/Ocean.kt
@@ -156,7 +156,7 @@ data class Ocean(
     val botDep: Float,
     val lon: Float,
     val lat: Float,
-    val dateT: Int
+    val dateT: Long
 )
 
 /**
